requirements.txt
streamlit
tavily-python
pypdf
python-dotenv
langchain
langchain-openai
langchain-community
redis
beautifulsoup4
langchain-chroma